remove casts

+++ examples/polybench_hls/adi.py
@@ -21,10 +21,10 @@ def adi_np(u, v, p, q, TSTEPS, N):
 
     a = int(-mul1 / 2)
     b = int(1 + mul1)
-    c = int(a)
-    d = int(-mul2 / 2)
-    e = int(1 + mul2)
-    f = int(d)
+    c = (a)
+    d = (-mul2 / 2)
+    e = (1 + mul2)
+    f = (d)
 
     for t in range(1, TSTEPS + 1):
         for i in range(1, N - 1):
@@ -71,10 +71,10 @@ def adi(ttype, TSTEPS, N):
 
     a = int(-mul1 / 2)
     b = int(1 + mul1)
-    c = int(a)
-    d = int(-mul2 / 2)
-    e = int(1 + mul2)
-    f = int(d)
+    c = (a)
+    d = (-mul2 / 2)
+    e = (1 + mul2)
+    f = (d)
 
     def kernel_adi[
         T: (int32, int32), TSTEPS: int32, N: int32
